remove positioning on Project components
Project components had position fixed on them, thereby causing them to stack on top of each other. Different Projects are now showing, but the carousel does not move/displace the slides sufficiently upon clicking the next/prev button. As a result, clicking the nav buttons results in minimal shift in the slides, and only 1.5 slides are visible.